requirements.txt
eventlet==0.23.0
flask-cors==3.0.3
gunicorn==19.7.1
